Added improvements to the project.

+++ src/controllers/deliveryAddress.controller.js
@@ -1,10 +1,11 @@
 import { response400, response500 } from '../helpers/responseHandlers'
 import DeliveryAddress from '../models/DeliveryAddress'
 import User from '../models/User'
+import Order from '../models/Order'
 
 const getDelideryAdressesByUser = async (req, res) => {
    try {
-      const { userId } = req.query
+      const userId = req.user.uid.toString()
 
       const deliveryAddress = await DeliveryAddress.find({ user: userId })
 
@@ -21,12 +22,16 @@ const getDelideryAdressesByUser = async (req, res) => {
 const addDeliveryAddress = async (req, res) => {
    try {
       const { deliveryAddress } = req.body
-      const { user: userId } = deliveryAddress
+
+      const userId = req.user.uid.toString()
 
       const user = await User.exists({ _id: userId })
       if (!user) return response400(res, 'No existe ningun usuario con ese ID')
 
-      const newDeliveryAddress = new DeliveryAddress(deliveryAddress)
+      const newDeliveryAddress = new DeliveryAddress({
+         ...deliveryAddress,
+         user: userId,
+      })
       const deliveryAddressSaved = await newDeliveryAddress.save()
 
       return res.status(200).json({
@@ -85,6 +90,16 @@ const deleteDeliveryAddress = async (req, res) => {
          return response400(
             res,
             'No existe ningun dirección de entrega con ese ID'
+         )
+
+      const existsOrder = await Order.findOne({
+         deliveryAddress: deliveryAddressesId,
+      }).lean()
+
+      if (existsOrder)
+         return response400(
+            res,
+            'La dirección de entraga esta agregada en varios pedidos, no se puede eliminar'
          )
 
       await DeliveryAddress.deleteOne({
